fix bad pull from display-rework
