+++ magpylib/_src/defaults/defaults_utility.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 from magpylib._src.defaults.defaults_values import DEFAULTS
 
 
-SUPPORTED_PLOTTING_BACKENDS = ("matplotlib", "plotly", "matplotlib_old")
+SUPPORTED_PLOTTING_BACKENDS = ("matplotlib", "plotly")
 
 
 SYMBOLS_MATPLOTLIB_TO_PLOTLY = {
