+++ magpylib/_src/display/plotly/plotly_sensor_mesh.py
