+++ magpylib/_src/display/plotly/plotly_utility.py
@@ -1,0 +1,147 @@
+"""utility functions for plotly backend"""
+import numpy as np
+
+
+def merge_mesh3d(*traces):
+    """Merges a list of plotly mesh3d dictionaries. The `i,j,k` index parameters need to cumulate
+    the indices of each object in order to point to the right vertices in the concatenated
+    vertices. `x,y,z,i,j,k` are mandatory fields, the `intensity` and `facecolor` parameters also
+    get concatenated if they are present in all objects. All other parameter found in the
+    dictionary keys are taken from the first object, other keys from further objects are ignored.
+    """
+    merged_trace = {}
+    L = np.array([0] + [len(b["x"]) for b in traces[:-1]]).cumsum()
+    for k in "ijk":
+        if k in traces[0]:
+            merged_trace[k] = np.hstack([b[k] + l for b, l in zip(traces, L)])
+    for k in "xyz":
+        merged_trace[k] = np.concatenate([b[k] for b in traces])
+    for k in ("intensity", "facecolor"):
+        if k in traces[0] and traces[0][k] is not None:
+            merged_trace[k] = np.hstack([b[k] for b in traces])
+    for k, v in traces[0].items():
+        if k not in merged_trace:
+            merged_trace[k] = v
+    return merged_trace
+
+
+def merge_scatter3d(*traces):
+    """Merges a list of plotly scatter3d. `x,y,z` are mandatory fields and are concatenated with a
+    `None` vertex to prevent line connection between objects to be concatenated. Keys are taken
+    from the first object, other keys from further objects are ignored.
+    """
+    merged_trace = {}
+    for k in "xyz":
+        merged_trace[k] = np.hstack([pts for b in traces for pts in [[None], b[k]]])
+    for k, v in traces[0].items():
+        if k not in merged_trace:
+            merged_trace[k] = v
+    return merged_trace
+
+
+def merge_traces(*traces):
+    """Merges a list of plotly 3d-traces. Supported trace types are `mesh3d` and `scatter3d`.
+    All traces have be of the same type when merging. Keys are taken from the first object, other
+    keys from further objects are ignored.
+    """
+    if len(traces) > 1:
+        if traces[0]["type"] == "mesh3d":
+            trace = merge_mesh3d(*traces)
+        elif traces[0]["type"] == "scatter3d":
+            trace = merge_scatter3d(*traces)
+    elif len(traces) == 1:
+        trace = traces[0]
+    else:
+        trace = []
+    return trace
+
+
+def getIntensity(vertices, axis) -> np.ndarray:
+    """Calculates the intensity values for vertices based on the distance of the vertices to
+    the mean vertices position in the provided axis direction. It can be used for plotting
+    fields on meshes. If `mag` See more infos here:https://plotly.com/python/3d-mesh/
+
+    Parameters
+    ----------
+    vertices : ndarray, shape (n,3)
+        The n vertices of the mesh object.
+    axis : ndarray, shape (3,)
+        Direction vector.
+
+    Returns
+    -------
+    Intensity values: ndarray, shape (n,)
+    """
+    p = np.array(vertices).T
+    pos = np.mean(p, axis=1)
+    m = np.array(axis)
+    intensity = (p[0] - pos[0]) * m[0] + (p[1] - pos[1]) * m[1] + (p[2] - pos[2]) * m[2]
+    # normalize to interval [0,1] (necessary for when merging mesh3d traces)
+    ptp = np.ptp(intensity)
+    ptp = ptp if ptp != 0 else 1
+    intensity = (intensity - np.min(intensity)) / ptp
+    return intensity
+
+
+def getColorscale(
+    color_transition=0,
+    color_north="#E71111",  # 'red'
+    color_middle="#DDDDDD",  # 'grey'
+    color_south="#00B050",  # 'green'
+) -> list:
+    """Provides the colorscale for a plotly mesh3d trace. The colorscale must be an array
+    containing arrays mapping a normalized value to an rgb, rgba, hex, hsl, hsv, or named
+    color string. At minimum, a mapping for the lowest (0) and highest (1) values is required.
+    For example, `[[0, 'rgb(0,0,255)'], [1,'rgb(255,0,0)']]`. In this case the colorscale
+    is created depending on the north/middle/south poles colors. If the middle color is
+    None, the colorscale will only have north and south pole colors.
+
+    Parameters
+    ----------
+    color_transition : float, default=0.1
+        A value between 0 and 1. Sets the smoothness of the color transitions from adjacent colors
+        visualization.
+    color_north : str, default=None
+        Magnetic north pole color.
+    color_middle : str, default=None
+        Color of area between south and north pole.
+    color_south : str, default=None
+        Magnetic north pole color.
+
+    Returns
+    -------
+    colorscale: list
+        Colorscale as list of tuples.
+    """
+    if color_middle is False:
+        colorscale = [
+            [0.0, color_south],
+            [0.5 * (1 - color_transition), color_south],
+            [0.5 * (1 + color_transition), color_north],
+            [1, color_north],
+        ]
+    else:
+        colorscale = [
+            [0.0, color_south],
+            [0.2 - 0.2 * (color_transition), color_south],
+            [0.2 + 0.3 * (color_transition), color_middle],
+            [0.8 - 0.3 * (color_transition), color_middle],
+            [0.8 + 0.2 * (color_transition), color_north],
+            [1.0, color_north],
+        ]
+    return colorscale
+
+
+def clean_legendgroups(fig):
+    """removes legend duplicates"""
+    frames = [fig.data]
+    if fig.frames:
+        data_list = [f["data"] for f in fig.frames]
+        frames.extend(data_list)
+    for f in frames:
+        legendgroups = []
+        for t in f:
+            if t.legendgroup not in legendgroups and t.legendgroup is not None:
+                legendgroups.append(t.legendgroup)
+            elif t.legendgroup is not None and t.legendgrouptitle.text is None:
+                t.showlegend = False
