+++ magpylib/_src/style.py
@@ -285,8 +285,7 @@ class Model3d(MagicProperties):
             pairs, or a callable returning the equivalent dictionary.
 
         backend: str
-            Plotting backend corresponding to the trace. Can be one of
-            `['generic', 'matplotlib', 'plotly']`.
+            Plotting backend corresponding to the trace. Can be one of `['matplotlib', 'plotly']`.
 
         constructor: str
             Model constructor function or method to be called to build a 3D-model object
@@ -329,8 +328,7 @@ class Trace3d(MagicProperties):
     Parameters
     ----------
     backend: str
-        Plotting backend corresponding to the trace. Can be one of
-        `['generic', 'matplotlib', 'plotly']`.
+        Plotting backend corresponding to the trace. Can be one of `['matplotlib', 'plotly']`.
 
     constructor: str
         Model constructor function or method to be called to build a 3D-model object
@@ -483,16 +481,14 @@ class Trace3d(MagicProperties):
 
     @property
     def backend(self):
-        """Plotting backend corresponding to the trace. Can be one of
-        `['generic', 'matplotlib', 'plotly']`."""
+        """Plotting backend corresponding to the trace. Can be one of `['matplotlib', 'plotly']`."""
         return self._backend
 
     @backend.setter
     def backend(self, val):
-        backends = ["generic"] + list(SUPPORTED_PLOTTING_BACKENDS)
-        assert val is None or val in backends, (
+        assert val is None or val in SUPPORTED_PLOTTING_BACKENDS, (
             f"The `backend` property of {type(self).__name__} must be one of"
-            f"{backends},\n"
+            f"{SUPPORTED_PLOTTING_BACKENDS},\n"
             f"but received {repr(val)} instead."
         )
         self._backend = val
