+++ tests/test_Coumpound_setters.py
@@ -8,7 +8,7 @@ import pytest
 from scipy.spatial.transform import Rotation as R
 
 import magpylib as magpy
-from magpylib._src.display.traces_base import make_Prism
+from magpylib._src.display.base_traces import make_Prism
 
 magpy.defaults.display.backend = "plotly"
