+++ tests/test_display_plotly.py
@@ -3,7 +3,7 @@ import plotly.graph_objects as go
 import pytest
 
 import magpylib as magpy
-from magpylib._src.display.traces_generic import get_generic_traces
+from magpylib._src.display.plotly.plotly_display import get_plotly_traces
 from magpylib._src.exceptions import MagpylibBadUserInput
 from magpylib.magnet import Cuboid
 from magpylib.magnet import Cylinder
@@ -162,6 +162,38 @@ def test_display_bad_style_kwargs():
     fig = go.Figure()
     with pytest.raises(ValueError):
         magpy.show(canvas=fig, markers=[(1, 2, 3)], style_bad_style_kwarg=None)
+
+
+def test_draw_unsupported_obj():
+    """test if a object which is not directly supported by magpylib can be plotted"""
+    magpy.defaults.display.backend = "plotly"
+
+    class UnkwnownNoPosition:
+        """Dummy Class"""
+
+    class Unkwnown1DPosition:
+        """Dummy Class"""
+
+        position = [0, 0, 0]
+
+    class Unkwnown2DPosition:
+        """Dummy Class"""
+
+        position = [[0, 0, 0]]
+        orientation = None
+
+    with pytest.raises(AttributeError):
+        get_plotly_traces(UnkwnownNoPosition())
+
+    traces = get_plotly_traces(Unkwnown1DPosition)
+    assert (
+        traces[0]["type"] == "scatter3d"
+    ), "make trace has failed, should be 'scatter3d'"
+
+    traces = get_plotly_traces(Unkwnown2DPosition)
+    assert (
+        traces[0]["type"] == "scatter3d"
+    ), "make trace has failed, should be 'scatter3d'"
 
 
 def test_extra_model3d():
